Distance is now used from user data

+++ app/src/main/java/com/varsitycollege/birdvue/RegisterActivity.kt
@@ -132,7 +132,7 @@ class RegisterActivity : AppCompatActivity() {
                     user!!.updateProfile(changeRequest)
 
                     //Create user object for user - default distance is 10
-                    writeNewUser(user.uid, name, email, 10, true)
+                    writeNewUser(user.uid, name, email, 10.0, true)
 
                     //Sign in immediately
                     auth.signInWithEmailAndPassword(email, password).addOnCompleteListener {
@@ -166,7 +166,7 @@ class RegisterActivity : AppCompatActivity() {
         }
     }
 
-    private fun writeNewUser(userId: String, name: String, email: String, maxDistance: Int, metric: Boolean) {
+    private fun writeNewUser(userId: String, name: String, email: String, maxDistance: Double, metric: Boolean) {
         val user = User(userId, name, email, maxDistance, metric)
         val ref = database.getReference("users")
         ref.child(userId).setValue(user).addOnSuccessListener{
