+++ app/src/main/java/com/varsitycollege/birdvue/data/User.kt
@@ -7,6 +7,6 @@ data class User(
     val id: String ?= null,
     val username: String ?= null,
     val email: String ?= null,
-    val maxDistance: Int ?= null,
+    val maxDistance: Double ?= null,
     val metricUnits: Boolean ?= null
 )
